Moved test reporter to test/support
TestReporter was for testing only and didn't belong in Elixometer's main module.
